move size-history library to caliptra-infra
